Various refactorings for points
