+++ core/src/main/java/com/github/skjolber/packing/packer/plain/PlainPackager.java
@@ -15,9 +15,9 @@ import com.github.skjolber.packing.api.StackPlacement;
 import com.github.skjolber.packing.api.StackValue;
 import com.github.skjolber.packing.api.Stackable;
 import com.github.skjolber.packing.api.ep.Point3D;
-import com.github.skjolber.packing.api.ep.XYPlanePoint3D;
 import com.github.skjolber.packing.deadline.PackagerInterruptSupplier;
 import com.github.skjolber.packing.ep.points3d.ExtremePoints3D;
+import com.github.skjolber.packing.ep.points3d.XYPlanePoint3D;
 import com.github.skjolber.packing.packer.AbstractPackagerBuilder;
 import com.github.skjolber.packing.packer.DefaultPackResult;
 import com.github.skjolber.packing.packer.DefaultPackResultComparator;
@@ -123,19 +123,30 @@ public class PlainPackager extends AbstractPlainPackager<Point3D<StackPlacement>
 							continue;
 						}
 
-						long pointSupport = calculateXYSupport(extremePoints3D, point3d, stackValue);
+						long pointSupport = -1; // cache for costly measurement
 						if(bestIndex != -1) {
 							Point3D<StackPlacement> bestPoint = extremePoints3D.getValue(bestPointIndex);
 							
-							if(!isBetter(bestStackable, bestPoint, bestStackValue, bestSupport, box, point3d, stackValue, pointSupport)) {
+							if(point3d.getMinZ() > bestPoint.getMinZ()) {
+								continue;
+							}
+							
+							pointSupport = calculateXYSupport(extremePoints3D, point3d, stackValue);
+							
+							if(pointSupport < bestSupport) {
+								continue;
+							}
+							
+							if(stackValue.getArea() <= bestStackValue.getArea()) {
 								continue;
 							}
 						}
+						
 						if(constraint != null && !constraint.supports(stack, box, stackValue, point3d.getMinX(), point3d.getMinY(), point3d.getMinZ())) {
 							continue;
 						}
 						
-						bestSupport = pointSupport;
+						bestSupport = pointSupport == -1 ? calculateXYSupport(extremePoints3D, point3d, stackValue) : pointSupport;
 						bestPointIndex = k;
 						bestIndex = i;
 						bestStackValue = stackValue;
@@ -176,54 +187,9 @@ public class PlainPackager extends AbstractPlainPackager<Point3D<StackPlacement>
 				stack, remainingStackables.isEmpty(), index);
 	}
 
-	protected boolean isBetter(Stackable referenceStackable, Point3D<StackPlacement> referencePoint, StackValue referenceStackValue, long referenceSupport, Stackable candidateBox, Point3D<StackPlacement> candidatePoint,
-			StackValue candidateStackValue, long candidateSupport) {
-		// ********************************************
-		// * Prefer lowest point
-		// ********************************************
-
-		// compare supported area
-		if(candidateSupport == referenceSupport) {
-
-			if(candidatePoint.getMinZ() == referencePoint.getMinZ()) {
-
-				if(candidateStackValue.getArea() == referenceStackValue.getArea()) {
-
-					// compare sideways support	
-					referenceSupport = referencePoint.calculateXZSupport(referenceStackValue.getDx(), referenceStackValue.getDz()) +
-							referencePoint.calculateYZSupport(referenceStackValue.getDy(), referenceStackValue.getDz());
-
-					candidateSupport = candidatePoint.calculateXZSupport(candidateStackValue.getDx(), candidateStackValue.getDz()) +
-							candidatePoint.calculateYZSupport(candidateStackValue.getDy(), candidateStackValue.getDz());
-
-					if(candidateSupport == referenceSupport) {
-						// if everything is equal, the point with the tightest fit 
-						return candidatePoint.getArea() < referencePoint.getArea();
-					}
-					return candidateSupport > referenceSupport;
-				}
-				return candidateStackValue.getArea() > referenceStackValue.getArea();
-
-			}
-			return candidatePoint.getMinZ() < referencePoint.getMinZ();
-		}
-		return candidateSupport > referenceSupport;
-	}
-	
 	protected long calculateXYSupport(ExtremePoints3D<StackPlacement> extremePoints3D, Point3D<StackPlacement> referencePoint, StackValue stackValue) {
-
-		Placement3D xyPlane = null;
-		
 		long sum = 0;
-		if(referencePoint instanceof XYPlanePoint3D) {
-			XYPlanePoint3D xyPlanePoint3D = (XYPlanePoint3D)referencePoint;
 
-			sum += referencePoint.calculateXYSupport(stackValue.getDx(), stackValue.getDy());
-			// discount the area covered 
-			
-			xyPlane = xyPlanePoint3D.getXYPlane();
-		}
-		
 		int minX = referencePoint.getMinX();
 		int minY = referencePoint.getMinY();
 		
@@ -236,9 +202,6 @@ public class PlainPackager extends AbstractPlainPackager<Point3D<StackPlacement>
 		
 		List<StackPlacement> placements = extremePoints3D.getPlacements();
 		for(StackPlacement stackPlacement : placements) {
-			if(stackPlacement == xyPlane) {
-				continue;
-			}
 			if(stackPlacement.getAbsoluteEndZ() == z) {
 				
 				// calculate the common area
