+++ extreme-points/src/main/java/com/github/skjolber/packing/ep/points2d/DefaultXYSupportPoint2D.java
@@ -4,8 +4,6 @@ import java.io.Serializable;
 
 import com.github.skjolber.packing.api.Placement2D;
 import com.github.skjolber.packing.api.ep.Point2D;
-import com.github.skjolber.packing.api.ep.XSupportPoint2D;
-import com.github.skjolber.packing.api.ep.YSupportPoint2D;
 
 public class DefaultXYSupportPoint2D<P extends Placement2D & Serializable> extends SimplePoint2D<P> implements XSupportPoint2D, YSupportPoint2D {
