+++ extreme-points/src/main/java/com/github/skjolber/packing/ep/points3d/DefaultYZPlanePoint3D.java
@@ -4,7 +4,6 @@ import java.io.Serializable;
 
 import com.github.skjolber.packing.api.Placement3D;
 import com.github.skjolber.packing.api.ep.Point3D;
-import com.github.skjolber.packing.api.ep.YZPlanePoint3D;
 
 public class DefaultYZPlanePoint3D<P extends Placement3D & Serializable> extends SimplePoint3D<P> implements YZPlanePoint3D {
