+++ extreme-points/src/main/java/com/github/skjolber/packing/ep/points3d/ExtremePoints3D.java
@@ -10,7 +10,6 @@ import org.eclipse.collections.impl.list.mutable.primitive.IntArrayList;
 import com.github.skjolber.packing.api.Placement3D;
 import com.github.skjolber.packing.api.ep.ExtremePoints;
 import com.github.skjolber.packing.api.ep.Point3D;
-import com.github.skjolber.packing.api.ep.YZPlanePoint3D;
 
 /**
  *
