+++ test/src/main/java/com/github/skjolber/packing/test/assertj/AbstractPoint3DAssert.java
@@ -3,9 +3,6 @@ package com.github.skjolber.packing.test.assertj;
 import org.assertj.core.api.AbstractObjectAssert;
 
 import com.github.skjolber.packing.api.ep.Point3D;
-import com.github.skjolber.packing.api.ep.XYPlanePoint3D;
-import com.github.skjolber.packing.api.ep.XZPlanePoint3D;
-import com.github.skjolber.packing.api.ep.YZPlanePoint3D;
 
 @SuppressWarnings("rawtypes")
 public abstract class AbstractPoint3DAssert<SELF extends AbstractPoint3DAssert<SELF, ACTUAL>, ACTUAL extends Point3D>
@@ -87,160 +84,6 @@ public abstract class AbstractPoint3DAssert<SELF extends AbstractPoint3DAssert<S
 		isNotNull();
 		if(actual.getMaxZ() != z) {
 			failWithMessage("Expected z " + z + ", not " + actual.getMaxZ() + " (was " + getCoordinates() + ")");
-		}
-		return myself;
-	}
-
-	public SELF isXYSupportAt(int x, int y) {
-		isNotNull();
-		if(!actual.isSupportedXYPlane(x, y)) {
-			if(actual instanceof XYPlanePoint3D) {
-				XYPlanePoint3D support = (XYPlanePoint3D)actual;
-
-				failWithMessage("Expected xy support at " + x + "x" + y + ", was " + support.getSupportedXYPlaneMaxX() + "x" + support.getSupportedXYPlaneMaxY() + " for " + actual);
-
-			} else {
-				failWithMessage("Expected xy support at " + x + "x" + y + ", was none for " + actual);
-			}
-		}
-		return myself;
-	}
-
-	public SELF isXZSupportAt(int x, int z) {
-		isNotNull();
-		if(!actual.isSupportedXZPlane(x, z)) {
-			if(actual instanceof XZPlanePoint3D) {
-				XZPlanePoint3D support = (XZPlanePoint3D)actual;
-
-				failWithMessage("Expected xz support at " + x + "x" + z + ", was " + support.getSupportedXZPlaneMaxX() + "x" + support.getSupportedXZPlaneMaxZ() + " for " + actual);
-			} else {
-				failWithMessage("Expected xz support at " + x + "x" + z + ", was none for " + actual);
-			}
-		}
-		return myself;
-	}
-
-	public SELF isYZSupportAt(int y, int z) {
-		isNotNull();
-		if(!actual.isSupportedYZPlane(y, z)) {
-			if(actual instanceof YZPlanePoint3D) {
-				YZPlanePoint3D support = (YZPlanePoint3D)actual;
-
-				failWithMessage("Expected yz support at " + y + "x" + z + ", was " + support.getSupportedYZPlaneMaxY() + "x" + support.getSupportedYZPlaneMaxZ() + " for " + actual);
-
-			} else {
-				failWithMessage("Expected yz support at " + y + "x" + z + ", was none for " + actual);
-			}
-		}
-		return myself;
-	}
-
-	public SELF isNoXYSupportAt(int x, int y) {
-		isNotNull();
-		if(actual.isSupportedXYPlane(x, y)) {
-			if(actual instanceof XYPlanePoint3D) {
-				XYPlanePoint3D support = (XYPlanePoint3D)actual;
-
-				failWithMessage("Expected no xy support at " + x + "x" + y + ", was " + support.getSupportedXYPlaneMaxX() + "x" + support.getSupportedXYPlaneMaxY() + " for " + actual);
-			}
-		}
-		return myself;
-	}
-
-	public SELF isNoXZSupportAt(int x, int z) {
-		isNotNull();
-		if(actual.isSupportedXZPlane(x, z)) {
-			if(actual instanceof XZPlanePoint3D) {
-				XZPlanePoint3D support = (XZPlanePoint3D)actual;
-
-				failWithMessage("Expected no xz support at " + x + "x" + z + ", was " + support.getSupportedXZPlaneMaxX() + "x" + support.getSupportedXZPlaneMaxZ() + " for " + actual);
-			}
-		}
-		return myself;
-	}
-
-	public SELF isNoYZSupportAt(int y, int z) {
-		isNotNull();
-		if(actual.isSupportedYZPlane(y, z)) {
-			if(actual instanceof YZPlanePoint3D) {
-				YZPlanePoint3D support = (YZPlanePoint3D)actual;
-
-				failWithMessage("Expected no yz support at " + y + "x" + z + ", was " + support.getSupportedYZPlaneMaxY() + "x" + support.getSupportedYZPlaneMaxZ() + " for " + actual);
-			}
-		}
-		return myself;
-	}
-
-	public SELF isMaxXYSupport(int x, int y) {
-		isNotNull();
-		if(actual instanceof XYPlanePoint3D) {
-			XYPlanePoint3D support = (XYPlanePoint3D)actual;
-
-			if(support.getSupportedXYPlaneMaxX() != x || support.getSupportedXYPlaneMaxY() != y) {
-				failWithMessage("Expected xy support at " + x + "x" + y + ", was " + support.getSupportedXYPlaneMaxX() + "x" + support.getSupportedXYPlaneMaxY() + " for " + actual);
-			}
-
-		} else {
-			failWithMessage("Expected xy support at " + x + "x" + y + ", was none for " + actual);
-		}
-		return myself;
-	}
-
-	public SELF isMaxXZSupport(int x, int z) {
-		isNotNull();
-		if(actual instanceof XZPlanePoint3D) {
-			XZPlanePoint3D support = (XZPlanePoint3D)actual;
-
-			if(support.getSupportedXZPlaneMaxX() != x || support.getSupportedXZPlaneMaxZ() != z) {
-				failWithMessage("Expected xz support at " + x + "x" + z + ", was " + support.getSupportedXZPlaneMaxX() + "x" + support.getSupportedXZPlaneMaxZ() + " for " + actual);
-			}
-		} else {
-			failWithMessage("Expected xz support at " + x + "x" + z + ", was none for " + actual);
-		}
-		return myself;
-	}
-
-	public SELF isMaxYZSupport(int y, int z) {
-		isNotNull();
-		if(actual instanceof YZPlanePoint3D) {
-			YZPlanePoint3D support = (YZPlanePoint3D)actual;
-
-			if(support.getSupportedYZPlaneMaxY() != y || support.getSupportedYZPlaneMaxZ() != z) {
-				failWithMessage("Expected yz support at " + y + "x" + z + ", was " + support.getSupportedYZPlaneMaxY() + "x" + support.getSupportedYZPlaneMaxZ() + " for " + actual);
-			}
-
-		} else {
-			failWithMessage("Expected yz support at " + y + "x" + z + ", was none for " + actual);
-		}
-		return myself;
-	}
-
-	public SELF isNoXYSupport() {
-		isNotNull();
-		if(actual instanceof XYPlanePoint3D) {
-			XYPlanePoint3D support = (XYPlanePoint3D)actual;
-
-			failWithMessage("Expected no xy support, was " + support.getSupportedXYPlaneMaxX() + "x" + support.getSupportedXYPlaneMaxY());
-		}
-		return myself;
-	}
-
-	public SELF isNoXZSupport() {
-		isNotNull();
-		if(actual instanceof XZPlanePoint3D) {
-			XZPlanePoint3D support = (XZPlanePoint3D)actual;
-
-			failWithMessage("Expected no xz support, was " + support.getSupportedXZPlaneMaxX() + "x" + support.getSupportedXZPlaneMaxZ());
-		}
-		return myself;
-	}
-
-	public SELF isNoYZSupport() {
-		isNotNull();
-		if(actual instanceof YZPlanePoint3D) {
-			YZPlanePoint3D support = (YZPlanePoint3D)actual;
-
-			failWithMessage("Expected no yz support, was " + support.getSupportedYZPlaneMaxY() + "x" + support.getSupportedYZPlaneMaxZ());
 		}
 		return myself;
 	}
